Update hid.c
wMaxPacketSize: The maximum packet size received or sent by this endpoint.
For synchronous endpoints, this value is used to indicate the bus time reserved by the host in the schedule. This is the time required for the data payload of each (micro) frame. The payload time is the bus time required to send one frame of data. In the actual data During the transmission process, the actual bandwidth used by the pipeline may be less than the reserved bandwidth. Think about it, if the actual used bandwidth is more than the reserved bandwidth, then you will lose the number;
For its type of endpoint, bit10~bit0 specifies the maximum packet size (in bytes);
Bit12~bit11 are valid for synchronization and interrupt endpoints of high-speed transmission: bit12~bit11 can specify the number of additional communications for each microframe. Everyone must know that in high-speed transmission, when a transaction times out, within a microframe time The number of retransmissions. If set to 00b (None), it means that only one transaction can be transmitted in a micro frame without additional timeout retransmission. If set to 01b, it means that two transactions can be transmitted in a micro frame. , there is an additional retransmission opportunity. As can be seen from the following, a microframe can have up to two opportunities to retransmit the transaction. If the microframe ends or fails, you need to wait until the next microframe to continue sending the transaction;